Ziko's Weather App, v1.0
Fully functional with pre-determined city, units and language. Please be
in the CET timezone to get a proper reading. ;)
A few things to do in order for app to be timezone agnostic and
give the user at least a (pre-determined) selection of cities.
Some forecast icons would've been nice as well.